feat: Add functions to shuffle and calculate distribution
Added functions shuffleGenerate and ObsExpChrOb to shuffle features and calculate observed and expected distribution of each chromosome. The functions take input factor data and genome information, with optional interval inclusion and exclusion. Shuffling is done based on a random seed for consistency. The ObsExpChrOb function allows for parallel processing if specified.